Move SimulatorActions from core to habitat_sim
