+++ examples/new_actions.py
@@ -17,9 +17,10 @@ import numpy as np
 
 import habitat
 import habitat_sim
-from habitat.sims.habitat_simulator.action_spaces import (
+from habitat.sims.habitat_simulator.actions import (
     HabitatSimV1ActionSpaceConfiguration,
 )
+from habitat.sims.habitat_simulator import SimulatorActions
 from habitat.tasks.nav.nav_task import SimulatorTaskAction
 
 
@@ -96,11 +97,11 @@ class NoNoiseStrafe(HabitatSimV1ActionSpaceConfiguration):
     def get(self):
         config = super().get()
 
-        config[habitat.SimulatorActions.STRAFE_LEFT] = habitat_sim.ActionSpec(
+        config[SimulatorActions.STRAFE_LEFT] = habitat_sim.ActionSpec(
             "noisy_strafe_left",
             NoisyStrafeActuationSpec(0.25, noise_amount=0.0),
         )
-        config[habitat.SimulatorActions.STRAFE_RIGHT] = habitat_sim.ActionSpec(
+        config[SimulatorActions.STRAFE_RIGHT] = habitat_sim.ActionSpec(
             "noisy_strafe_right",
             NoisyStrafeActuationSpec(0.25, noise_amount=0.0),
         )
@@ -113,11 +114,11 @@ class NoiseStrafe(HabitatSimV1ActionSpaceConfiguration):
     def get(self):
         config = super().get()
 
-        config[habitat.SimulatorActions.STRAFE_LEFT] = habitat_sim.ActionSpec(
+        config[SimulatorActions.STRAFE_LEFT] = habitat_sim.ActionSpec(
             "noisy_strafe_left",
             NoisyStrafeActuationSpec(0.25, noise_amount=0.05),
         )
-        config[habitat.SimulatorActions.STRAFE_RIGHT] = habitat_sim.ActionSpec(
+        config[SimulatorActions.STRAFE_RIGHT] = habitat_sim.ActionSpec(
             "noisy_strafe_right",
             NoisyStrafeActuationSpec(0.25, noise_amount=0.05),
         )
@@ -131,7 +132,7 @@ class StrafeLeft(SimulatorTaskAction):
         return "strafe_left"
 
     def step(self, *args, **kwargs):
-        return self._sim.step(habitat.SimulatorActions.STRAFE_LEFT)
+        return self._sim.step(SimulatorActions.STRAFE_LEFT)
 
 
 @habitat.registry.register_task_action
@@ -140,12 +141,12 @@ class StrafeRight(SimulatorTaskAction):
         return "strafe_right"
 
     def step(self, *args, **kwargs):
-        return self._sim.step(habitat.SimulatorActions.STRAFE_RIGHT)
+        return self._sim.step(SimulatorActions.STRAFE_RIGHT)
 
 
 def main():
-    habitat.SimulatorActions.extend_action_space("STRAFE_LEFT")
-    habitat.SimulatorActions.extend_action_space("STRAFE_RIGHT")
+    SimulatorActions.extend_action_space("STRAFE_LEFT")
+    SimulatorActions.extend_action_space("STRAFE_RIGHT")
 
     config = habitat.get_config(config_paths="configs/tasks/pointnav.yaml")
     config.defrost()
