+++ habitat/sims/habitat_simulator/__init__.py
@@ -1,3 +1,5 @@
+from habitat.sims.habitat_simulator.actions import SimulatorActions
+
 from habitat.core.registry import registry
 from habitat.core.simulator import Simulator
 
@@ -13,7 +15,7 @@ def _try_register_habitat_sim():
 
     if has_habitat_sim:
         from habitat.sims.habitat_simulator.habitat_simulator import HabitatSim
-        from habitat.sims.habitat_simulator.action_spaces import (
+        from habitat.sims.habitat_simulator.actions import (
             HabitatSimV1ActionSpaceConfiguration,
         )
     else:
