+++ test/test_habitat_sim.py
@@ -10,7 +10,7 @@ import os
 import numpy as np
 import pytest
 
-from habitat import SimulatorActions
+from habitat.sims.habitat_simulator import SimulatorActions
 from habitat.config.default import get_config
 from habitat.sims import make_sim
